+++ NEWS.md
@@ -1,4 +1,5 @@
 # teal.modules.clinical 0.8.9.9000
+* New module `tm_t_exposure` was added for the analysis of duration of exposure for risk management plan. 
 * Corrected `tm_a_mmrm` to be able to consider the treatment variable in additional interactions. 
 * New module `tm_t_shift_by_grade` was added for the analysis of grade laboratory abnormalities. 
 * Fixed `tm_t_binary_outcome` and `tm_t_rsp` so that correct CI estimation method is chosen for Proportions Difference in Stratified Analysis (i.e. Wald-type confidence interval with CMH weights).